# openapi_check.json
{"openapi":"3.1.0","info":{"title":"Evo KokoroTTS","description":"API local de Text-to-Speech multilíngue com Kokoro-82M e Edge TTS. Suporta catálogo de idiomas, vozes por engine, geração completa em MP3/WAV, streaming em MP3 e ajuste de pitch para vozes Edge, incluindo presets infantis.","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Verificar status da API","description":"Retorna o status atual da API, o dispositivo em uso e se o pipeline padrão do Kokoro já foi carregado.","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/languages":{"get":{"summary":"Listar idiomas disponíveis","description":"Retorna todos os idiomas disponíveis na instalação atual, incluindo idiomas atendidos por Kokoro e Edge TTS.","operationId":"list_languages_languages_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/LanguageInfo"},"type":"array","title":"Response List Languages Languages Get"}}}}}}},"/voices":{"get":{"summary":"Listar vozes disponíveis","description":"Retorna o catálogo completo de vozes, incluindo engine, idioma e presets sugeridos de speed e pitch.","operationId":"list_voices_voices_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/VoiceInfo"},"type":"array","title":"Response List Voices Voices Get"}}}}}}},"/tts":{"post":{"summary":"Gerar áudio completo","description":"Sintetiza o texto inteiro e retorna um arquivo MP3 ou WAV. Para vozes Edge, speed=1.0 e pitch=0 aplicam os presets default da voz quando definidos.","operationId":"text_to_speech_tts_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TTSRequest"}}},"required":true},"responses":{"200":{"description":"Áudio gerado com sucesso.","content":{"application/json":{"schema":{}}}},"400":{"description":"Parâmetros inválidos ou voz indisponível."},"500":{"description":"Falha interna durante a geração."},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tts/stream":{"post":{"summary":"Gerar áudio por streaming","description":"Sintetiza e retorna o áudio em fluxo MP3. Para vozes Edge, speed=1.0 e pitch=0 aplicam os presets default da voz quando definidos.","operationId":"text_to_speech_stream_tts_stream_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TTSStreamRequest"}}},"required":true},"responses":{"200":{"description":"Stream de áudio iniciado com sucesso.","content":{"application/json":{"schema":{}}}},"400":{"description":"Parâmetros inválidos ou voz indisponível."},"500":{"description":"Falha interna durante a geração."},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status","description":"Estado da API: ok ou loading."},"device":{"type":"string","title":"Device","description":"Dispositivo atual de inferência: cuda ou cpu."},"model_loaded":{"type":"boolean","title":"Model Loaded","description":"Indica se o pipeline Kokoro padrão já foi carregado."}},"type":"object","required":["status","device","model_loaded"],"title":"HealthResponse"},"LanguageInfo":{"properties":{"id":{"type":"string","title":"Id","description":"ID normalizado do idioma."},"name":{"type":"string","title":"Name","description":"Nome nativo do idioma."},"label":{"type":"string","title":"Label","description":"Rótulo amigável para UI."},"kokoro_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Kokoro Code","description":"Código de idioma usado pelo Kokoro, quando aplicável."},"supported":{"type":"boolean","title":"Supported","description":"Indica se o idioma está disponível nesta instalação."},"sample_text":{"type":"string","title":"Sample Text","description":"Texto de exemplo para teste rápido."}},"type":"object","required":["id","name","label","supported","sample_text"],"title":"LanguageInfo"},"TTSRequest":{"properties":{"text":{"type":"string","maxLength":10000,"minLength":1,"title":"Text","description":"Texto para sintetizar.","examples":["Olá! Este é um teste de síntese de voz infantil em português."]},"voice":{"type":"string","title":"Voice","description":"ID da voz. Use GET /voices para listar todas as opções.","default":"pf_dora","examples":["pt_f_menina"]},"speed":{"type":"number","maximum":2.0,"minimum":0.5,"title":"Speed","description":"Velocidade da fala. Para vozes Edge, 1.0 usa o preset default da voz quando existir.","default":1.0,"examples":[1.0]},"pitch":{"type":"integer","maximum":80.0,"minimum":-80.0,"title":"Pitch","description":"Ajuste de pitch em Hz. Para vozes Edge, 0 usa o preset default da voz quando existir.","default":0,"examples":[46]},"format":{"type":"string","enum":["mp3","wav"],"title":"Format","description":"Formato de saída.","default":"mp3","examples":["mp3"]}},"type":"object","required":["text"],"title":"TTSRequest"},"TTSStreamRequest":{"properties":{"text":{"type":"string","maxLength":10000,"minLength":1,"title":"Text","description":"Texto para sintetizar em streaming.","examples":["Texto longo para playback progressivo em tempo real."]},"voice":{"type":"string","title":"Voice","description":"ID da voz. Use GET /voices para listar todas as opções.","default":"pf_dora","examples":["pt_m_menino"]},"speed":{"type":"number","maximum":2.0,"minimum":0.5,"title":"Speed","description":"Velocidade da fala. Para vozes Edge, 1.0 usa o preset default da voz quando existir.","default":1.0,"examples":[1.05]},"pitch":{"type":"integer","maximum":80.0,"minimum":-80.0,"title":"Pitch","description":"Ajuste de pitch em Hz. Para vozes Edge, 0 usa o preset default da voz quando existir.","default":0,"examples":[28]}},"type":"object","required":["text"],"title":"TTSStreamRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VoiceInfo":{"properties":{"id":{"type":"string","title":"Id","description":"ID da voz."},"name":{"type":"string","title":"Name","description":"Nome curto exibido na UI."},"gender":{"type":"string","title":"Gender","description":"Gênero da voz."},"lang":{"type":"string","title":"Lang","description":"Idioma da voz."},"lang_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Lang Code","description":"Código de idioma usado pelo Kokoro, quando aplicável."},"engine":{"type":"string","title":"Engine","description":"Engine responsável pela síntese: kokoro ou edge."},"default_speed":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Default Speed","description":"Preset sugerido de velocidade para a voz."},"default_pitch":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Default Pitch","description":"Preset sugerido de pitch em Hz para a voz."},"description":{"type":"string","title":"Description","description":"Descrição curta da voz."}},"type":"object","required":["id","name","gender","lang","engine","description"],"title":"VoiceInfo"}}}}